refactor: main is divided into module
