+++ src/main.rs
@@ -1,45 +1,11 @@
-use std::{env, fmt};
 use std::fmt::Formatter;
 use std::io::BufReader;
-use chrono::Utc;
+use std::{env, fmt};
 use tokio_cron_scheduler::{Job, JobScheduler};
 
-fn play_sound(file: &std::fs::File) -> Result<(), ProcessError>{
-    let file = file.try_clone().map_err(|_| ProcessError::FileReadError)?;
-    let decoder = rodio::Decoder::new(BufReader::new(file)).map_err(|_| ProcessError::FileReadError)?;
-    // OutputStream (_stream) is needed to play sound.
-    let (_stream, handle) = rodio::OutputStream::try_default().map_err(|_| ProcessError::OutputStreamError)?;
-    let sink = rodio::Sink::try_new(&handle).map_err(|_| ProcessError::OutputStreamError)?;
+mod service;
 
-    sink.append(decoder);
-    sink.sleep_until_end();
-    Ok(())
-}
-
-struct JobCreateService {
-}
-
-impl JobCreateService {
-    fn new() -> Self {
-        Self {}
-    }
-
-    /// Create a job to add scheduler
-    ///
-    /// * `expression` - Text to represent the schedule.
-    /// * `filename` - Sound file path. (String is used to move the ownership to closure and save file path within job to return)
-    fn create_sound_job(expression: &str, file_path: String) -> Job {
-        Job::new(expression, move |_uuid, _l| {
-            println!("{:?}", Utc::now());
-            let file = std::fs::File::open(file_path.clone()).unwrap();
-            match play_sound(&file) {
-                Err(err) => println!("{}", err),
-                _ => {},
-            }
-            println!("{:?}", Utc::now());
-        }).unwrap()
-    }
-}
+use service::job_creation_service::JobCreateService;
 
 struct JobExecuteService {
     scheduler: JobScheduler,
@@ -81,19 +47,5 @@ async fn main() {
 
     loop {
         tokio::time::sleep(core::time::Duration::from_millis(500)).await;
-    }
-}
-
-enum ProcessError {
-    FileReadError,
-    OutputStreamError,
-}
-
-impl fmt::Display for ProcessError {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        match *self {
-            ProcessError::FileReadError => write!(f, "file state is invalid"),
-            ProcessError::OutputStreamError => write!(f, "output stream setting is failed"),
-        }
     }
 }
